replace LookBackFilter to separate file
